apps/mission-control/api/requirements.txt
fastapi>=0.115.0
httpx>=0.28.1
numpy>=2.1.0
pydantic>=2.8.0
pytest>=8.3.5
pytest-asyncio>=1.0.0
pyyaml>=6.0.2
stable-baselines3>=2.5.0
torch>=2.4.0
uvicorn[standard]>=0.30.0
